gptel-anthropic: Remove auto-cache when sending media
* gptel-anthropic.el (gptel--anthropic-parse-multipart): Don't
impose caching automatically when including media in Anthropic-API
requests.  Now that `gptel-cache' is available, it is left up to
the user to turn on caching for Anthropic models if they wish to.

Caching unconditionally here also leads to errors: when you add a
cache point to the system prompt, messages array, tools AND after
this image, the API complains that there are too many cache
markers.