gptel-openai: Fix JSON encoding for function args when parsing
Tool call arguments are supplied to and by OpenAI as a JSON
object.  When reading tool call args/results from the buffer,
gptel ends up JSON encoding the args twice: once to get results as
a JSON object, and once again when sending it.  Non-ascii UTF-8
chars are encoded to codepoints, and attempting to JSON-encode
these codepoints again causes the request to fail.  Fix by
decoding the args to UTF-8 after the first encoding.
(#1001 and #1004)

* gptel-openai.el (gptel--parse-list, gptel--parse-buffer): Decode
the JSON encoded tool args (as read from a chat buffer) to UTF-8.