gptel-curl: Pass payload unencoded to Curl process
gptel fails on MS-windows 10/11 in Japanese environments (#1033)
because the built-in Curl does not accept UTF-8 encoded
arguments (and expects CP932).  Fix by decoding the JSON-encoded
payload (which always uses UTF-8) to whatever is appropriate to
the user's process-environment.  This fix was provided by
@shingo256.

Ensure that we don't needlessly create a copy of the payload (if
it can be avoided) by setting the NOCOPY flag when running
`decode-coding-string'.

* gptel-curl.el (gptel-curl--get-args): Use `decode-coding-string'
around the JSON-encoded payload.

* gptel.el (gptel--url-get-response): Do NOT use
`decode-coding-string'.  This change slipped in by mistake in
90b41b1.  It is not required here since the string never leaves
Emacs' environment.  Further, decoding the UTF-8 encoded JSON and
attempting to send non-ASCII chars actually causes an encoding
error.