gptel-curl: Improve thinking-block handling in response content
* gptel-curl.el (gptel-curl--stream-filter, gptel-curl--sentinel):
The <think> tag in the response can appear with or without
newlines after and before the block.  Handle all these cases
correctly for streaming and no-streaming responses.   Previously
it was assumed that <think> could be preceded by spaces but not
newlines. (#1028)

We still assume that no non-whitespace text can appear before a
<think> tag in a response.  If it does the thinking text will not
be recognized as such.

* gptel.el (gptel--url-get-response): Ditto for the url-retrieve
path.