gptel-openai: Improve reasoning_content handling in stream parser
* gptel-openai.el (gptel-curl--parse-stream): Experimental change.
If the streaming response includes a blank (only whitespace)
content field, do not assume that the reasoning block has ended or
that it doesn't exist.  This is to work around a bug in vLLM's
treatment of the GLM-4.5 model, and I intend to revert it after a
couple of months.  (#1028)

* test: Add tests for llama.cpp and vLLM stream parsing using
output from the GLM-4.5 model.