gptel-gh: Change how tokens are managed
Removed the file based persistent storage of the tokens, as they were
stored in clear text. They will only be stored in variables.

The initial value of the github-token can now be set by providing a
function that provides it. This enables this token to be securely
stored outside of Emacs.