gptel-gh: Change how tokens are managed
Introduced overloadable function variables that determines how a token
is loaded and saved.

gptel-gh: Introduced logic for backends to share the github-token

This commit introduces proper support for more than one GitHub
account. By stating which account a backend should be considered to be
connected to, it is possible to ensure that we don't e.g. login using
a personal account to a backend considered for the work account.

The file based persistence will make use of the stated GitHub username
and make that a suffix to the configured file path. As the username
for GitHub accounts may only contain [0-9a-zA-Z], they can safely be
used for filenames.

The custom save/load function now takes the username as a parameter
so that it can decide on how to proceed.