gptel: Extend save/restore state to support tools (#963)
- Extend saved chat state to include gptel tools.  Active tool
names are written to and read from Org properties (in Org mode)
and file-local variables (elsewhere).

- Additionally, handle "deletions" correctly.  That is, remove
existing Org properties and file-local variables for settings that
are set to nil.  For some settings (like the backend/model) a nil
value does not make sense and is ignored.  The system message
could be nil and this edge case is not yet handled.  Other
persisted settings are handled correctly.

* gptel-org.el (gptel-org--send-with-props)
(gptel-org--entry-properties, gptel-org--restore-state)
(gptel-org-set-properties): Handle the GPTEL_TOOLS property.

* gptel.el (gptel--tool-names, gptel--backend-name)
(gptel--restore-state, gptel--save-state): Use a new variable
`gptel--tool-names' to persist gptel tool names for the chat
session.