gptel-request: Remove gptel--wrap-user-prompt
`gptel--wrap-user-prompt' is a vestige of an older context
wrapping implementation in gptel.

When `gptel-use-context' is set to 'user', we used to construct
the full messages array, and then concat the context into the last
user prompt in this array.

Post 0.9.7, gptel pastes the context text into the prompt
construction buffer instead, and lets the usual
`gptel--parse-buffer' routines handle constructing the messages
array with the context included.

So `gptel--wrap-user-prompt' has been superfluous for a while,
except for an edge case: It is used to include media (binary)
files added to gptel's global context (via `gptel-add-file', etc),
and these media files were always included with the first user
prompt in the array.

We do this even if `gptel-use-context' is set to system, since not
all backends uniformly support including documents/images in the
system prompt.

This use is regrettably still unavoidable, because we don't have a
convenient way of including an image in the prompt construction
buffer that isn't major-mode specific (such as via Org links).

In keeping with its only remaining use, rename
`gptel--wrap-user-prompt' to `gptel--inject-media' in all backend
implementations, simplify it and remove the unused code branch in
the function.

* gptel-anthropic.el (gptel--wrap-user-prompt):
(gptel--inject-media):
* gptel-bedrock.el (gptel--wrap-user-prompt):
(gptel-bedrock--inject-media-context):
(gptel--inject-media):
(gptel-bedrock--inject-text-context):
* gptel-gemini.el (gptel--wrap-user-prompt):
(gptel--inject-media):
* gptel-kagi.el (gptel--wrap-user-prompt):
(gptel-make-kagi):
* gptel-ollama.el (gptel--wrap-user-prompt):
(gptel--inject-media):
* gptel-openai.el (gptel--wrap-user-prompt):
(gptel--inject-media):
* gptel-request.el (gptel--realize-query):
(gptel--wrap-user-prompt):