gptel-context: Change how context is specified
A common complaint abotu gptel is that the context handling does
not have a good API for programmatic usage, or for buffer-local
context management (see #459, #475, #481, #548, #984 and several
more).  To make the context handling more ergonomic, we introduce
several changes to `gptel-context--alist' in this commit.

The main advantage is that context can now be added by pushing
files and buffers (objects, not buffer names) to this alist.  This
variable can be made buffer local, etc.  The full API will be
described in the next commit, after renaming
`gptel-context--alist' to a user option.

This commit fixes the plumbing required for this change,
`gptel-context--wrap' is now a pure function -- it returns a
sanitized version of `gptel-context--alist' that can be fed to the
context wrapper (`gptel-context-string-function').  Since
`gptel-context--alist' is intended to be a user option, we modify
it less frequently than before to avoid user surprise.
Unfortunately we still need to do this a few times for gptel's
menus to function as intended.

* gptel-context.el (gptel-context-string-function):
(gptel-context-remove):
(gptel-context-remove-all):
(gptel-context--collect-media):
(gptel-context--collect):
(gptel-context--insert-buffer-string):
(gptel-context--string):
(gptel-context--buffer-setup):
(gptel-context-confirm):

* gptel-transient.el (gptel--describe-infix-context): Don't
run `gptel-context--wrap' when showing `gptel-menu'.  Use
the (possibly out of date) gptel-context--alist here.

* gptel.el (gptel-mode): Adjust header-line display.

The effect of this staleness is yet to be gauged.