gptel: Handle deletions correctly when inheriting stickiness
* gptel.el (gptel--inherit-stickiness): When inheriting stickiness
in gptel-mode buffer changes, handle deletions correctly.

Previously, we checked if the "pre" argument of the after-change
function was zero, assuming that this covers all text changes
except deletions.  However, the buffer can be modified by
primitive functions like `replace-buffer-contents' or
`replace-match' where new text is inserted and old text is removed
in a single action, in which case "pre" will be nonzero.

To exclude only pure deletions, we now check if the "beg" and
"end" of the change region differ from each other.