gptel-transient: Add command to read and apply preset
* gptel-transient.el (gptel--read-apply-preset, gptel--preset):
The gptel presets menu does not scale well and cannot handle more
than a few presets -- a known problem.  This is a design problem
that will have to be addressed in the future.  For now, add an
escape hatch to read and apply a preset using `completing-read'
with "@ @", where "@" brings up the menu.