gptel-rewrite: Use exit-minibuffer instead of quitting
* gptel-rewrite.el (gptel--rewrite-read-message): Use
`exit-minibuffer' instead of `minibuffer-quit-recursive-edit'.

The latter causes a quit to be raised, leading to the
disconcerting `ring-bell-function' action for actions that aren't
quits, like starting a rewrite or switching to the rewrite
transient menu.  Using `exit-minibuffer' and exiting `gptel-rewrite'
cleanly is fine because the inverted flow control in this function
runs the user's intended action using timers anyway.