gptel: Add declarative preset composition (#905)
When using presets, it is often desirable to modify the current
value of a gptel option/variable instead of replacing it
wholesale.  For example, it is common to want to add more LLM
tools via a preset to an existing set in gptel-tools.  To this
end, add a small, declarative DSL for use in gptel preset
definitions.  You can now do the following:

  (gptel-make-preset 'websearch
    :tools '(:append ("search_web" "read_url")))

to add to the current list in `gptel-tools' instead of replacing
it.  Another example is adding to a system prompt.  Continuing the
same example, we can add to the current system message when
including web search tools:

  (gptel-make-preset 'websearch
    :tools  '(:append ("search_web" "read_url"))
    :system '(:prepend "Use the provided tools to search the web
                        for up-to-date information.\n\n"))

The currently supported operators and data types are :append and
:prepend (for lists or strings) and :merge (for plists).  Support
for more types and operators will be added in the future.

* gptel.el (gptel--modify-value, gptel--apply-preset)
(gptel-make-preset): Add a declarative way to modify a list or
string, via `gptel--modify-value'.  This function takes an
original value and a plist spec declaring how to modify it (the
"modification spec").  It returns the changed value, or the new
value if it is not a valid modification spec.  It is
non-destructive.

Add support for this to `gptel--apply-preset', taking care to
handle some incompatible type combinations, like appending a
string to a function.  Not all incompatibilities are handled yet.

Document new features in `gptel-make-preset'.

NEWS: Document new feature.