gptel: Improve preset validation, modify-spec handling
* gptel.el (gptel--modify-value):  Use keywordp instead of
checking for it the hard way.  I keep forgetting that keywordp
exists.  (Like car-safe, this is a weird choice of functions to
implement in C.)

* gptel-transient.el (gptel--preset-mismatch-p,
gptel--read-apply-preset): Set `gptel--preset' when applying a
preset via `completing-read'.

Try to handle modify-specs for :tools when checking if there is a
preset mismatch.  Handle both tools and tool names in the :tools
spec correctly. (#1099)

Handling all possible modify-specs correctly in
`gptel--preset-mismatch-p' is impossible, or at least impractical.
For example, if a spec contains

  :system (:function 'do-something)

where do-something is not idempotent, then we have nothing to
compare the current system message (which is

  (do-something gptel--system-message))

against gptel--system-message.