gptel-transient: Fix appending to nil system message
Handle correctly the case of appending an additional system
instruction ("d" in `gptel-menu') to a nil system message.  nil is
a list so dispatching on the list type is a bug.

* gptel-transient.el (gptel--merge-additional-directive): Use
`cons' instead of `list' on cl-typecase to avoid matching a `nil'
system message, which would then give an error.