gptel-org: Add link validation to Org mode gptel buffers
A commonly reported complaint about gptel-mode is that it's
difficult to see why a link will or will not be sent when
`gptel-track-media' is turned on.  This commit adds some of the
functionality required to annotate a link to be sent via `jit-lock'.

Org mode links are being worked on first, with support for
Markdown coming in the future.

* gptel-org.el (gptel-org--link-regex): Regex covering recognized
Org link syntax, extracted from a function into a constant.
(gptel--parse-media-links): Use this regex.
(gptel-org--annotate-links): New function to annotate links that will
be sent.  Not yet hooked up to jit-lock.