gptel-org: Make Org link validation comprehensive
* gptel-org.el (gptel-org--annotate-links): Make link validation
more comprehensive by checking for everything about a link -- we
want to ensure that a link will be sent under the current gptel
configuration iff jit-lock indicates that it will.

This means doing all the link checks: file readability, binary
file check, mimetype support, url capability and so on inside
jit-lock.  This check is in the process of being optimized.

* gptel-request.el (gptel--link-type-cache): Cache to store
determination of binary status of files.  This is required
because:

1. We read 512 bytes from the file in `gptel--file-binary-p' to
determine if the file is binary-encoded.
2. When validating links, this check runs via jit-lock, and can
thus run too many times.  In general it's a bad idea to do file
access inside jit-lock.
3. To get around this we cache the binary status of files we've
seen.