gptel: Add code block folding to Markdown (#845)
Because markdown-mode doesn't include it for some reason.  This
makes using reasoning/tools in markdown-mode buffers much more
pleasant.

Fold tool and reasoning blocks by default in markdown-mode
buffers.

* gptel.el (gptel-markdown-cycle-block): New command for cycling
GFM-style code blocks in Markdown.

(gptel--markdown-block-map): A keymap to bind the cycling command
to text as a text-property.  It is added to the opening and
closing lines of a code block.

(gptel-mode): The `keymap' text property is part of
`font-lock-extra-managed-props' in Markdown, so we cannot set it
as a text-property.  Instead, add a font-lock keyword to Markdown
buffers when enabling gptel-mode.

(gptel--insert-response, gptel--display-reasoning-stream,
gptel--display-tool-results): Fold tool and reasoning blocks in
non-Org mode buffers when insering them.  Add the keymap property
to the opening and closing code fences, so that folding/unfolding
the block via `gptel-markdown-cycle-block' is bound to Tab.  Note
that this method doesn't work in markdown-mode itself because of
the above reason, but it should work in text-mode or other
buffers.

* NEWS (New features and UI changes): Mention new folding feature.