gptel: Add minor mode to highlight LLM responses
New, oft-requested feature `gptel-highlight-mode' to highlight
LLM response text regions (as well as tool calls and ignored
regions).

* gptel.el (gptel-highlight-fringe): Bitmap for fringe
highlighting.

(gptel-highlight-mode): New minor-mode to highlight gptel LLM
responses.  Works in any buffer in Emacs.

(gptel-highlight-methods): New user option to configure how
highlighting should be performed.  The default is to use the left
margin, since this works on TTY also.  Other options are a fringe
bitmap and a response face.  The face, while comprehensive, can
obscure other Org/Markdown formatting and is thus not the default.

(gptel-response-fringe-highlight, gptel-response-highlight): Faces
for highlighting.

NOTE: The face names are subject to change, since it is not clear
yet if it might be better to suffix faces with "-face".  gptel's
approach here is currently inconsistent and will be fixed
eventually.

(gptel-highlight--update, gptel-highlight--decorate,
gptel-highlight--fringe-prefix, gptel-highlight--margin-prefix):
Implementation of `gptel-highlight-mode'.

* README.org (Usage, FAQ): Mention `gptel-highlight-mode'.
* NEWS (New features and UI changes): Mention change.