gptel-context.el: move the context before the prompt
Enhance the KV cache reuse further. This is particularly helpful with
the gptel-rewrite-directives-hook that may change the prompt at every
request depending on the language of the buffer.

The model seem to follow the prompt better when it's at the end (at
least open weight qwen3 coder instruct), for example it starts to
follow the instruction not add markdown fences around the rewritten
code that is otherwise ignored on a long context.

Supposedly the fact the context format is actively using the markdown
quotes, washes the directive not to use the markdown quotes if it's
done at the top of the system message.