gptel: Make :append actions on strings idempotent
* gptel.el (gptel--modify-value): A common use of the :append and
:prepend keys when specifying string-valued preset keys is to add
to the system message.  If a preset is applied twice, this
concatenation happens twice as well, which is undesirable more
than it is intentional.  Make this action idempotent, so presets
can be applied multiple times without repeatedly appending
strings.

In general, list values are still appended as specified without
any checks.  (When specifying :tools in particular, the list is
deduplicated before sending a request.)

This change is experimental.