gptel-context: Fix context-buffer for new context spec
* gptel-context.el (gptel-context--insert-buffer-string): Fix bug
with nconc-ing nil.

(gptel-context--buffer-setup): Fix for `gptel-context's plist
spec.  Only `:overlays' is handled so far, bounds/line ranges in
files/buffers are ignored.  This needs to be fixed in the future.

* gptel-transient.el (gptel--describe-infix-context): Fix the
context chunk count for `gptel-context's new plist spec.  Only
`:overlays' is handled so far, bounds/line ranges are ingored.
This needs to be fixed as well.