gptel-context: Move the context before the prompt (#1110)
gptel-context.el (gptel-context--wrap-in-buffer): Place the
context chunks before the system message instead of after.

The objective of this change is to enhance KV cache reuse, based
on the following empirical observations:

1. Context chunks are large; the rest of the system prompt is
small (on average).
2. The instructions change more often than the context does in a
   conversation (or rewrite/refactoring task).

So we want the move the smaller, more frequently changing
component further down in the chain.

Observation by Andrea Arcangeli (@aagit):

"As a bonus, the model seem to follow the prompt better when it's
at the end (at least open weight qwen3 coder instruct), for
example it starts to follow the instruction not add markdown
fences around the rewritten code that is otherwise ignored on a
long context.

Supposedly the fact the context format is actively using the
markdown quotes, washes the directive not to use the markdown
quotes if it's done at the top of the system message."

-----

This change is experimental, and if we see regressions in LLMs'
tendency to follow instructions, it may be reverted.