gptel-context: Sanitize stale context, fix context removal
* gptel-context.el (gptel-context-remove): Fix for
`gptel-context's new plist spec.

(gptel-context--buffer-setup): Fix for new plist spec.

* gptel-transient.el (gptel-menu): Sanitize `gptel-context' before
displaying the menu, clearing context from killed buffers.
Because `gptel-context--collect' is now pure and no longer sets
`gptel-context', we have to sanitize `gptel-context' _somewhere_!
gptel-menu might not be the best place for it, but at least this
way you don't see wrong counts in the context indicator in the
menu.