gptel-request: Set backend consistently in multi-turn requests
Some backend slots like the header and url can be functions that
use gptel-backend and gptel-model from their environment.
Previously, their "environment" was just the request buffer.  The
backend and model can change mid-request in multi-turn requests,
such as when employing sub-agents as tools, or if the user just
happens to change them when a request is in progress.  This caused
the request to throw an error as the wrong backend url/header
functions were called.

Hermetically "seal" the backend and model into the request INFO
plist, and let-bind these values whenever calling a function whose
behavior depends on them.  Currently this let-binding is required
only when dispatching a new turn, in `gptel--url-get-response' and
`gptel-curl--get-args', but this is something to keep in mind.

Whenever possible, use (plist-get info :backend) and:(plist-get
info :model) instead of the defvars `gptel-backend' and
`gptel-model'.

* gptel-request.el (gptel--realize-query): Store the model for the
request in INFO.
(gptel--url-get-response, gptel-curl--get-args): Let-bind the
backend and model around calls correctly.