gptel-ollama: Reinitialize :reasoning-block after tool calls
gptel-ollama.el (gptel-curl--parse-stream): Some models can reason
in phases, for instance:

- Reasoning block before a tool call
- Tool call
- New reasoning block after the tool call

This is different from calling tools as part of the reasoning
phase.

Reinitialize reasoning-block to nil when there is no reasoning.
This change is experimental, as this initialization will occur
every turn in the stream parser when there is no reasoning
content.