gptel: Refine gptel-send FSM handlers
Minor tweaks to the FSM handlers for `gptel-send'.  These changes
are made in the interest of more robust UI and state tracking, as
required by UI-heavy tools like the ones used by gptel-agent.

Details below.

* gptel.el (gptel-send--handlers): Add a new handler
`gptel--update-abort' for the ABRT
state, replacing the previous one.

(gptel--handle-post-insert, gptel--handle-error,
gptel--update-abort):

Introduce a new per-request post-response-functions list, as
:post in INFO.  It serves the same purpose as
`gptel-post-response-functions', but can be tailored individually
to each gptel-send request by tools or other FSM handlers.  It is
planned to be used for state cleanup, the same as
post-response-functions.  This change is a long-time coming.

NOTE: This is currently for internal use and the implementation is
subject to change.

TODO: Move the org-conversion and other post-response functions to
the :post hook to avoid having to match against active requests,
and move `gptel-post-response-functions' to gptel from
gptel-request.

Run the post-command hooks before pulsing the response region, as
the region to be pulsed can change.  Update the header-line last,
as it might be further modified by tools.

(gptel--update-tool-call, gptel--update-tool-ask): Update
`gptel--fsm-last' before tools run, not after.  This is important
for synchronous tool calls that need to examine it to fetch buffer
state.