gptel-anthropic: Allow multi-part system prompt
* gptel-anthropic.el (gptel--request-data): Allow system prompts
for Anthropic models to be multi-part, although all of them are
still required to be text.  This will be extended to other APIs
too, but it's not clear right now which OpenAI-compatible APIs
support multi-part system prompts and which don't.

For Anthropic, a multi-part system prompt can be specified as

  (("part1" "part2" "part3" ...))

Note the double-listing -- this is required since a list of
strings is interpreted as a system prompt followed by alternating
user and assistant messages:

  ("system" "user1" "llm1" "user2" "llm2" ...)

(gptel--parse-list, gptel--parse-buffer): When `gptel-cache' is
enabled, don't use the 1 hour cache TTL for Anthropic -- the
caching prices are considerably higher.  Stick to the default
ephemeral cache with a TTL of 5 min, this covers the most common
case of requests with extended tool use.