gptel: Use in-buffer tool call previews instead of overlays
Change how tool confirmation prompts work in gptel's default
buffer UI.  Previously the confirmation prompt was an overlay
after-string, now text is physically inserted into the buffer.

The main reason for the change is to show more details previews
for tool calls inline in the buffer, such as diffs.  Showing large
amounts of text as overlay after-strings does not work.  The plan
is to allow editing the tool call arguments inline, as well.

This also presents several disadvantages, such as font-lock
confusion and tool call confirmations in read-only buffers, so the
change is experimental.  An alternative where tool call arguments
can be inspected and edited in a separate buffer is being
considered as well.

* gptel.el (gptel--tool-preview-alist): New registry for tools to
add custom previewers that gptel will call to preview the tool
call.  This is not implemented as a `:preview' slot of the tool
definition because previewing is UI-specific behavior and the same
tool call can be previewed in different ways in different
gptel-based interfaces.

(gptel-tool-call-actions-map): New keymap for actions on pending
tool calls, extracted from an earlier inline definition.

(gptel--display-tool-calls): Implement in-buffer tool call previews.

(gptel--format-tool-call): Use font-lock-face instead of face, as
the preview now appears in font-locked buffers.

(gptel--accept-tool-calls, gptel--reject-tool-calls): Clean-up
tool call previews by deleting text and calling any tool-specific
previewer clean-up functions.