gptel-rewrite: Handle tool calls in gptel-rewrite
* gptel-rewrite.el (gptel--rewrite-directive-default): Add
additional instruction about not producing intermediate text.

(gptel--rewrite-callback): Handle tool calls in the middle of a
gptel-rewrite command.  The most common use for this is to include
a file/buffer-read tool to allow the LLM to fetch more context
before doing the rewrite.

The way this works is the following:

1. Tool calls requiring confirmation are displayed the usual way,
using the in-buffer overlays or the minibuffer if the buffer is
read-only.

2. Tool call results are ignored.

3. Text is inserted into the temp buffer as it is received, but
every time there is a tool call, all the text up to that point is
deleted.  Only the text generated after the final tool call is
retained in the overlay.

(gptel--rewrite-handlers): Dedicated FSM handler table for
gptel-rewrite.  We mainly need this for `gptel--update-tool-call',
which records the request in progress into `gptel--fsm-last'.  In
turn, this is needed to be able to call sub-agents or more
specialized tools.

(gptel--suffix-rewrite): Use the new handlers.