gptel-request: Evaluate system messages in request buffer
* gptel-request.el (gptel-request): Parse the system message,
evaluating function-valued system messages in the process, in the
request buffer instead of the temporary prompt buffer used to
construct the prompt.  (#1095, #1153)

This is necessary for correctness, since the system message
function is intended by the user to be evaluated in their current
environment.

The parsed system message (as returned by gptel--parse-directive)
is always a list, even when the original directive is a string.
This can cause issues when applying presets in the prompt that
modify the system message by appending to it.  To avoid this,
ensure that string-valued system prompts remain string-valued
until the query is finally realized in `gptel--realize-query'.

This latter fix is ugly, but the alternative is to make
`gptel--parse-directive' a polymorphic, sum-type function, which
appears to be uglier on the whole considering its many uses in
gptel's request pipeline and UI.

* NEWS (New features and UI changes): Mention bug fix.