gptel-request: Change post-request cleanup handlers API
Change the experimental per-request :post handler API for
`gptel-request'.  Now each function in :post is passed the INFO
plist as an argument.  This API is still internal/experimental,
since it's not yet clear what the best argument is.

Use the per-request :post API for deleting temporary JSON payload
files written to by `gptel-request'.  This is cleaner than
accumulating these deletions in `gptel-post-response-functions',
which runs only after gptel-send and is not request-specific.

For the same reasons, deleting the Org conversion buffers also
needs to happen in the :post hook, and will be moved over from
`gptel-post-response-functions' in the future.

* gptel-request.el (gptel-request--handlers, gptel--handle-post,
gptel-curl--get-args): Use the :post functions list to handle temp
file deletion.  Add transition handlers to
`gptel-request--transitions' to ensure that the :post hook always
runs.

* gptel.el (gptel--handle-post-insert, gptel--handle-error,
gptel--handle-abort): Change the calls to the :post functions to
pass them the INFO plist.