gptel-request: Run post-response after inserting prompt prefix
* gptel.el (gptel--handle-post-insert): Insert the prompt
prefix (if any) in gptel buffers before running
`gptel-post-response-functions'.  This is the correct order
because functions in this hook might depend on the presence of the
prompt prefix for their logic.  Using `gptel-end-of-response' is
one example. (#1163)

* gptel-request.el (gptel-post-response-functions): We want the
response pulsing to occur after the post-response hooks have run
-- since we moved running the hook to the end, remove the
hardcoded call to `pulse-momentary-highlight-region' in the
post-insert handler, and add it to `gptel-post-response-functions'
instead.  As a bonus, this can now be disabled.