gptel-context: Maintain context order in context buffer
gptel-context.el (gptel-context-confirm): When confirming changes
in the gptel context buffer, retain the context order instead of
reversing it.

`gptel-context--collect' returns a list in an order that's the
reverse of `gptel-context' -- this change was made in d8f3f67c to
ensure that new context is added at the end in the request payload
to maximize KV-cache reuse.  This commit adjusts the
context-buffer behavior to account for this change.