gptel: Improve end-of-response navigation
* gptel.el (gptel-beginning-of-response, gptel-end-of-response):
These commands accept both beg/end bounds arguments and an
additional numeric prefix arg.  The former are provided so they
can be used in the post-response hook, and the latter is for
interactive use.  Change these commands so that they prefer to use
the beg and end arguments when provided, taking precedence over
the text-property search that is otherwise employed. (#1163)