gptel-openai: Fix tool call handling when reasoning
* gptel-openai.el (gptel-curl--parse-stream,
gptel--parse-response): Some models (like deepseek-reasoner) call
tools when reasoning.  When sending the tool results, they expect
the reasoning content they supplied to be sent back as well.
Capture reasoning content and inject it to the messages array
along with the tool call. (#1176)

Reasoning content in streaming responses is now captured in two
places in the INFO plist (:reasoning and :reasoning-chunks), where
the former is consumed by the stream filter, as before.  The
latter is persistent until the end of the turn and used.

When parsing non-streaming responses, capture reasoning text
even if there is no other response text.  This is expected when a
response consists only of reasoning text and a tool call.

* gptel-request.el (gptel-curl--sentinel): Similarly, call the
callback with the reasoning text even when there is no other
response text in a non-streaming response.

* NEWS (Notable bug fixes): Mention fix.